// src/BLL/Models/SymbolMeta.model.ts
export default class SymbolMeta {
    id: string;
    t1mRsi14PrerequisitesCalculated: Boolean;
    t3mRsi14PrerequisitesCalculated: Boolean;
    t5mRsi14PrerequisitesCalculated: Boolean;
    t15mRsi14PrerequisitesCalculated: Boolean;
    t30mRsi14PrerequisitesCalculated: Boolean;
    t1hRsi14PrerequisitesCalculated: Boolean;
    t2hRsi14PrerequisitesCalculated: Boolean;
    t4hRsi14PrerequisitesCalculated: Boolean;
    t6hRsi14PrerequisitesCalculated: Boolean;
    t8hRsi14PrerequisitesCalculated: Boolean;
    t12hRsi14PrerequisitesCalculated: Boolean;
    t1dRsi14PrerequisitesCalculated: Boolean;
    t3dRsi14PrerequisitesCalculated: Boolean;
    t1wRsi14PrerequisitesCalculated: Boolean;
    t1MRsi14PrerequisitesCalculated: Boolean;
    symbolId: string;
}